Added an RmdOperator class
This introduces a third container along with postgres and airflow, which is called rserver. The RmdOperator sshs into this server to execute the Rmd. This also allows each operator to be an arbitrary file with a YAML frontmatter.

The rserver itself builds on the rocker/tidyverse docker image, and includes a DESCRIPTION file where additional packages can be added.